Split front and admin functionality into separate handlers
The plugin class now only runs a the list of handlers.
